gl4/glBlitFramebuffer: Correct an error condition
The original extension specification which defines this behavior states
that GL_INVALID_OPERATION is generated when SAMPLE_BUFFERS for EITHER
the read framebuffer or draw framebuffer is greater than zero. However,
the reference page states that GL_INVALID_OPERATION is generated when
SAMPLE_BUFFERS for BOTH framebuffers is greater than zero, conflicting
with the specification.